add first object
